Add repository name check in create form
GitHub's API returns 422 when trying to create
a repository with a name already existing in the
other repos. Detecting this code in the resource
now will raise a ResourceException which is in
addition handled in HackathonCreateController,
which will redirect to the form itself with the
valid values entered by the user passad as request
attributes.